backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
coreapi==2.0.8
django-crispy-forms==1.6.1
django-filter==0.15.3
Django==1.9.8
DjangoRestFramework==3.4.7
itypes==1.1.0             # via coreapi
Pillow==3.3.1
psycopg2==2.6.2
requests==2.11.1          # via coreapi
uritemplate==3.0.0        # via coreapi
